Update CMakeLists to modern practices
Modern CMake allows target's properties to be imported recursively.
This repository was using a combination of CMake targets and explicit
path references, which is undesirable.
It now uses targets for all dependencies, and the dependencies config
is now in a seperate file for manageability.

TODO: Testing and update CMake minimum version.